At the highlighted point, add a new RateLimitedSampler which implements GetSamplingIntent and can be used with the composable sampler framework in lib.rs. The sampler will compute an average Add a new RateLimitedSampler which implements GetSamplingIntent and can be used with the composable sampler framework in lib.rs. Here is an explanation of the functionality of this sampler: https://github.com/open-telemetry/opentelemetry-specification/blob/main/oteps/trace/0250-Composite_Samplers.md#consistentratelimiting
Do not modify the ComposableSampler enum, instead create a new implementation of GetSamplingIntent. We don't want this as a built-in, because there are potentially many good ways to rate limit. Here is an explanation of the functionality of this sampler:

ConsistentRateLimiting ConsistentRateLimiting is a composite sampler that helps control the average rate of sampled spans while allowing another sampler (the delegate) to provide sampling hints.

Required Arguments for ConsistentRateLimiting Composable (delegate) maximum sampling (throughput) target rate The sampler SHOULD measure and keep the average rate of incoming spans, and therefore also of the desired ratio between the incoming span rate to the target span rate. Upon invocation of its GetSamplingIntent operation, the composite sampler MUST get the SamplingIntent from the delegate sampler, passing the same arguments as received.

The returned SamplingIntent is constructed as follows.

If using the obtained threshold value as the final threshold would entail sampling more spans than the declared target rate, the sampler SHOULD set the threshold to a value that would meet the target rate. Several algorithms can be used for threshold adjustment, no particular behavior is prescribed by the specification though. The IsAdjustedCountReliable returns the result of calling this function on the SamplingIntent provided by the delegate. The GetAttributes function returns the result of calling this function on the SamplingIntent provided by the delegate. The UpdateTraceState function returns the Tracestate as returned by calling UpdateTraceState from the delegate's SamplingIntent.